powermind-front/requirements.txt
nicegui
supabase
python-dotenv
pydantic
email-validator